refactor: `ShareButton` as own component
