+++ packages/frontend/src/components/Modals/MarkerModal/MarkerModal.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback } from 'react'
+import React from 'react'
 import { useTranslation } from 'react-i18next'
 
 import { Report } from 'src/utils/types'
@@ -8,7 +8,8 @@ import { useStationsAndLines } from '../../../contexts/StationsAndLinesContext'
 import Skeleton, { useSkeleton } from '../../Miscellaneous/LoadingPlaceholder/Skeleton'
 import { useStationReports } from '../../../hooks/useStationReports'
 import { useStationDistance } from '../../../hooks/useStationDistance'
-import { sendAnalyticsEvent } from 'src/utils/analytics'
+
+import ShareButton from '../../Miscellaneous/ShareButton/ShareButton'
 
 import './MarkerModal.css'
 
@@ -41,48 +42,13 @@ const MarkerModal: React.FC<MarkerModalProps> = ({ className, children, selected
     const elapsedTimeMessage = useElapsedTimeMessage(elapsedTimeInMinutes, selectedMarker.isHistoric)
     const stationDistanceMessage = useStationDistanceMessage(stationDistance)
 
-    const handleShare = useCallback(
-        async (e: React.MouseEvent) => {
-            e.preventDefault()
-            try {
-                const directionTextShare = direction?.name || '?'
-                const lineTextShare = line || '?'
-
-                if (navigator.share) {
-                    await navigator.share({
-                        title: t('Share.title'),
-                        text: t('Share.text', {
-                            station: station.name,
-                            direction: directionTextShare,
-                            line: lineTextShare,
-                        }),
-                        url: window.location.href,
-                    })
-                    await sendAnalyticsEvent('Marker Shared', {
-                        meta: {
-                            station: station.name,
-                            line: line,
-                            direction: direction?.name,
-                        },
-                    })
-                } else {
-                    await navigator.clipboard.writeText(window.location.href)
-                    alert(t('Share.copied', 'Link copied to clipboard!'))
-                }
-            } catch (error) {
-                console.error('Error sharing:', error)
-            }
-        },
-        [t, station, line, direction]
-    )
-
     return (
         <div className={`marker-modal info-popup modal ${className}`}>
             {children}
             <h1>{station.name}</h1>
             <h2>
                 {line && (
-                    <span className="line-label" style={{ backgroundColor: getLineColor(line) }}>
+                    <span className="line" style={{ backgroundColor: getLineColor(line) }}>
                         {line}
                     </span>
                 )}
@@ -105,10 +71,7 @@ const MarkerModal: React.FC<MarkerModalProps> = ({ className, children, selected
                     <span className="disclaimer">{t('MarkerModal.inviteText')}</span>
                 </div>
                 {selectedMarker.message && <p className="description">{selectedMarker.message}</p>}
-                <button onClick={handleShare} className="share-button">
-                    <img src={process.env.PUBLIC_URL + '/icons/share-svgrepo-com.svg'} alt="Share" />
-                    <span>{t('Share.button')}</span>
-                </button>
+                <ShareButton report={selectedMarker} />
             </div>
         </div>
     )
